Separate out api vs function tests
